Move `Compiler` from `collector` binary to `toolchain`
